fix
внесла правки.
