+++ pom.xml
@@ -9,8 +9,8 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
         <mockito.version>3.12.4</mockito.version>
         <junit.version>4.13.2</junit.version>
         <jacoco-maven-plugin>0.8.7</jacoco-maven-plugin>
@@ -70,8 +70,8 @@
             <artifactId>maven-compiler-plugin</artifactId>
             <version>3.8.1</version>
             <configuration>
-                <source>11</source>
-                <target>11</target>
+                <source>17</source>
+                <target>17</target>
             </configuration>
         </plugin>
